--- src/amuse/community/bhtree/tests/test_bhtree_pytest.py
@@ -1,804 +1,0 @@
-
-# -*- coding: utf-8 -*-
-from typing import OrderedDict
-from amusetest import TestWithMPI
-import os
-import sys
-import numpy
-import time
-import math
-
-from amuse.community.bhtree.interface import BHTreeInterface, BHTree
-from amuse.support.exceptions import AmuseException
-from amuse.support.core import OrderedDictionary
-from amuse.units import constants
-from amuse.units import nbody_system
-from amuse.units import units
-from amuse.units import quantities
-from amuse import datamodel
-from amuse.datamodel import particle_attributes
-from amuse.ic import plummer
-try:
-    from matplotlib import pyplot
-    HAS_MATPLOTLIB = True
-except ImportError:
-    HAS_MATPLOTLIB = False
-
-
-from pytest import fixture
-from pytest import approx
-
-import pytest
-
-from amuse.units.quantities import none
-from amuse.units.quantities import to_quantity
-from amuse.units.quantities import is_quantity
-
-
-def check_comparable(x, y):
-    if is_quantity(x):
-        if not is_quantity(y) and not x.unit.base == none.base:
-            raise TypeError("Cannot compare quantity: {0} with non-quantity: {1}.".format(x, y))
-    elif is_quantity(y):
-        if not y.unit.base == none.base:
-            raise TypeError("Cannot compare non-quantity: {0} with quantity: {1}.".format(x, y))
-
-
-def convert_to_numeric(x, y, in_units):
-    if in_units:
-        return (x.value_in(in_units), y.value_in(in_units))
-    elif is_quantity(x) or is_quantity(y):
-        return (
-                to_quantity(x).value_in(to_quantity(y).unit),
-                to_quantity(y).value_in(to_quantity(y).unit)
-                )
-    else:
-        return (x, y)
-
-
-# helper function to check almost equal
-# NOTE: this is currently written for a scalar; the original
-# functions work with np.arrays
-# Also, we'll have to add functionality for the units here
-# TODO: add regression tests for new checks and old checks?
-def check_equal_with_abstol(x, y, digits, msg=""):
-    """Ported from failUnlessAlmostEqual."""
-    check_comparable(x, y)
-    x_num, y_num = convert_to_numeric(x, y, in_units=None)
-    assert x_num == approx(y_num, abs=10**(-digits)), msg
-
-
-def check_equal_with_reltol(x, y, digits=8, msg="", in_units=None):
-    """Ported from failUnlessAlmostRelativeEqual."""
-    check_comparable(x, y)
-    x_num, y_num = convert_to_numeric(x, y, in_units=in_units)
-    assert x_num == approx(y_num, rel=10**(-digits)), msg
-
-
-# TODO: rename the above with units? or find better name in general
-def check_equal_units(x, y, msg="", in_units=None):
-    """Ported from failUnlessEqual."""
-    check_comparable(x, y)
-    x_num, y_num = convert_to_numeric(x, y, in_units)
-    assert x_num == y_num, msg
-
-
-@fixture
-def bhtree_msun():
-    convert_nbody = nbody_system.nbody_to_si(1.0 | units.MSun, 149.5e6 | units.km) # for test1
-    instance = BHTree(convert_nbody)
-    instance.parameters.epsilon_squared = 0.001 | units.AU**2
-    instance.commit_parameters()
-
-    stars = datamodel.Stars(2)
-
-    sun = stars[0]
-    sun.mass = units.MSun(1.0)
-    sun.position = [0.0, 0.0, 0.0] | units.m
-    sun.velocity = [0.0, 0.0, 0.0] | units.ms
-    sun.radius = units.RSun(1.0)
-
-    earth = stars[1]
-    earth.mass = units.kg(5.9736e24)
-    earth.radius = units.km(6371)
-    earth.position = [149.5e6, 0.0, 0.0] | units.km
-    earth.velocity = [0.0, 29800, 0.0] | units.ms
-
-    instance.particles.add_particles(stars)
-    #instance.commit_particles() # this disables adding particles later on
-    yield stars, instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-@fixture
-def bhtree_kg(): # for test4, test6, test7, test11, test12, test13
-    convert_nbody = nbody_system.nbody_to_si(5.0 | units.kg, 10.0 | units.m)
-    instance = BHTree(convert_nbody)
-    instance.commit_parameters()
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-@fixture
-def bhtree_empty():
-    instance = BHTree()
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-@fixture
-def bhtree_test9():
-    instance = BHTree()
-    instance.initialize_code()
-    instance.parameters.epsilon_squared = 0.00001 | nbody_system.length**2
-
-    particles = datamodel.Particles(2)
-    particles.mass = [1.0, 1.0] | nbody_system.mass
-    particles.radius = [0.0001, 0.0001] | nbody_system.length
-    particles.position = [[0.0, 0.0, 0.0], [2.0, 0.0, 0.0]] | nbody_system.length
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | nbody_system.speed
-    instance.particles.add_particles(particles)
-
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-@fixture
-def bhtree_test10():
-    instance = BHTree()
-    instance.initialize_code()
-    instance.parameters.epsilon_squared = 0.00001 | nbody_system.length**2
-    instance.commit_parameters()
-
-    particles = datamodel.Particles(6)
-    particles.mass = 1.0 | nbody_system.mass
-    particles.radius = 0.00001 | nbody_system.length
-    particles.position = [[-1.0, 0.0, 0.0], [1.0, 0.0, 0.0], [0.0, -1.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, -1.0], [0.0, 0.0, 1.0]] | nbody_system.length
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | nbody_system.speed
-    instance.particles.add_particles(particles)
-    instance.commit_particles()
-
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-@fixture
-def bhtree_test14():
-    convert_nbody = nbody_system.nbody_to_si(1.0 | units.yr, 1.0 | units.AU)
-    instance = BHTree(convert_nbody)
-    instance.commit_parameters()
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-@fixture
-def bhtree_for_epsilon_squared_test():
-    convert_nbody = nbody_system.nbody_to_si(1.0 | units.MSun, 1.0 | units.AU)
-
-    particles = datamodel.Particles(2)
-    sun = particles[0]
-    sun.mass = 1.0 | units.MSun
-    sun.position = [0.0, 0.0, 0.0] | units.AU
-    sun.velocity = [0.0, 0.0, 0.0] | units.AU / units.yr
-    sun.radius = 1.0 | units.RSun
-
-    earth = particles[1]
-    earth.mass = 5.9736e24 | units.kg
-    earth.radius = 6371.0 | units.km
-    earth.position = [0.0, 1.0, 0.0] | units.AU
-    earth.velocity = [2.0*numpy.pi, -0.0001, 0.0] | units.AU / units.yr
-
-    initial_direction = math.atan((earth.velocity[0]/earth.velocity[1]))
-    final_direction = []
-    for log_eps2 in range(-9, 10, 2):
-        instance = BHTree(convert_nbody)
-        instance.initialize_code()
-        instance.parameters.epsilon_squared = 10.0**log_eps2 | units.AU ** 2
-        instance.particles.add_particles(particles)
-        instance.commit_particles()
-        instance.evolve_model(0.25 | units.yr)
-        final_direction.append(math.atan((instance.particles[1].velocity[0] /
-            instance.particles[1].velocity[1])))
-        #instance.stop()
-
-    yield instance, initial_direction, final_direction
-
-    instance.cleanup_code()
-    instance.stop()
-
-@fixture
-def bhtree_to_test_energy():
-    numpy.random.seed(0)
-    number_of_stars = 2
-    stars = plummer.new_plummer_model(number_of_stars)
-    stars.radius = 0.00001 | nbody_system.length
-    stars.scale_to_standard()
-
-    instance = BHTree()
-    instance.initialize_code()
-    instance.parameters.epsilon_squared = (1.0 / 20.0 / (number_of_stars**0.33333) | nbody_system.length)**2
-    instance.parameters.timestep = 0.004 | nbody_system.time
-    instance.parameters.timestep = 0.00001 | nbody_system.time
-    instance.commit_parameters()
-    instance.particles.add_particles(stars)
-    instance.commit_particles()
-
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-@fixture
-def bhtree_test18(): # used in test 18, 19, 20, 21, 22
-    particles = datamodel.Particles(2)
-    particles.x = [0.0, 10.0] | nbody_system.length
-    particles.y = 0 | nbody_system.length
-    particles.z = 0 | nbody_system.length
-    particles.radius = 0.005 | nbody_system.length
-    particles.vx = 0 | nbody_system.speed
-    particles.vy = 0 | nbody_system.speed
-    particles.vz = 0 | nbody_system.speed
-    particles.mass = 1.0 | nbody_system.mass
-
-    instance = BHTree()
-    instance.initialize_code()
-
-    # NOTE: the below was used in tests 18, 19, 20. but tests still pass with this turned off.
-    #instance.parameters.epsilon_squared = (0.01 | nbody_system.length)**2
-
-    instance.particles.add_particles(particles)
-
-    yield instance
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-@fixture
-def bhtree_test23():
-
-    particles = datamodel.Particles(2)
-    particles.x = [0.0, 10.0] | nbody_system.length
-    particles.y = 0.0 | nbody_system.length
-    particles.z = 0.0 | nbody_system.length
-    particles.vx = 1.0 | nbody_system.speed # this is different
-    particles.vy = 0.0 | nbody_system.speed
-    particles.vz = 0.0 | nbody_system.speed
-    particles.mass = 0.1 | nbody_system.mass # this is different
-
-    instance = BHTree(redirection="none")
-    instance.particles.add_particles(particles)
-    instance.commit_particles()
-
-
-    yield instance, particles
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-@fixture
-def bhtree_for_collision_detection():
-    particles = datamodel.Particles(7)
-    particles.mass = 0.001 | nbody_system.mass
-    particles.radius = 0.01 | nbody_system.length
-    particles.x = [-101.0, -100.0, -0.5, 0.5, 100.0, 101.0, 104.0] | nbody_system.length
-    particles.y = 0 | nbody_system.length
-    particles.z = 0 | nbody_system.length
-    particles.velocity = [[2, 0, 0], [-2, 0, 0]]*3 + [[-4, 0, 0]] | nbody_system.speed
-
-    instance = BHTree(redirection='none')
-    instance.initialize_code()
-    instance.parameters.set_defaults()
-
-    # TODO: make this an argument to the fixture?
-    # Uncommenting any of the following two lines will suppress collision detection
-# ~        instance.parameters.use_self_gravity = 0
-# ~        instance.parameters.epsilon_squared = 0.0 | nbody_system.length**2
-
-    instance.parameters.opening_angle = 0.1
-    instance.particles.add_particles(particles)
-    collisions = instance.stopping_conditions.collision_detection
-    collisions.enable()
-    instance.evolve_model(1.0 | nbody_system.time)
-
-    yield instance, particles, collisions
-
-    instance.cleanup_code()
-    instance.stop()
-
-
-def test_test1(bhtree_msun):
-    stars, bhtree = bhtree_msun
-    sun, earth = stars
-    postion_at_start = earth.position.value_in(units.AU)[0]
-
-    bhtree.evolve_model(365.0 | units.day)
-    bhtree.particles.copy_values_of_all_attributes_to(stars)
-
-    postion_after_full_rotation = earth.position.value_in(units.AU)[0]
-    check_equal_with_abstol(postion_at_start, postion_after_full_rotation, 3)
-
-    bhtree.evolve_model(365.0 + (365.0 / 2) | units.day)
-    bhtree.particles.copy_values_of_all_attributes_to(stars)
-    postion_after_half_a_rotation = earth.position.value_in(units.AU)[0]
-
-    check_equal_with_abstol(-postion_at_start, postion_after_half_a_rotation, 2)
-
-    bhtree.evolve_model(365.0 + (365.0 / 2) + (365.0 / 4) | units.day)
-    bhtree.particles.copy_values_of_all_attributes_to(stars)
-    postion_after_half_a_rotation = earth.position.value_in(units.AU)[1]
-
-    check_equal_with_abstol(-postion_at_start, postion_after_half_a_rotation, 1)
-
-
-def test_test4(bhtree_kg):
-    bhtree = bhtree_kg
-
-    index = bhtree.new_particle(
-        15.0 | units.kg,
-        10.0 | units.m, 20.0 | units.m, 30.0 | units.m,
-        0.0 | units.m/units.s, 0.0 | units.m/units.s, 0.0 | units.m/units.s,
-        10.0 | units.m
-    )
-    bhtree.commit_particles()
-    check_equal_units(bhtree.get_mass(index), 15.0 | units.kg, "new particle not added correctly")
-    check_equal_units(bhtree.get_radius(index), 10.0 | units.m)
-
-
-def test_test5(bhtree_empty):
-    instance = bhtree_empty
-    instance.commit_parameters() # TODO: can we move this away?
-
-    index = instance.new_particle(
-        15.0 | nbody_system.mass,
-        10.0 | nbody_system.length, 20.0 | nbody_system.length, 30.0 | nbody_system.length,
-        1.0 | nbody_system.speed, 1.0 | nbody_system.speed, 3.0 | nbody_system.speed,
-        10.0 | nbody_system.length
-    )
-    instance.commit_particles()
-    check_equal_units(instance.get_radius(index), 10.0 | nbody_system.length)
-    check_equal_units(instance.get_mass(index), 15.0 | nbody_system.mass)
-
-
-
-def test_test6(bhtree_kg):
-    instance = bhtree_kg
-    indices = instance.new_particle(
-        [15.0, 30.0] | units.kg,
-        [10.0, 20.0] | units.m, [20.0, 40.0] | units.m, [30.0, 50.0] | units.m,
-        # 1.0 | units.m/units.s, 1.0 | units.m/units.s, 3.0 | units.m/units.s
-        [0.0, 0.01] | units.m/units.s, [0.0, 0.01] | units.m/units.s, [0.0, 0.01] | units.m/units.s,
-        [10.0, 20.0] | units.m
-    )
-    instance.commit_particles()
-
-    check_equal_units(instance.get_mass(indices[0]), 15.0 | units.kg)
-
-    with pytest.raises(AmuseException) as excinfo:
-        instance.get_mass([4, 5])
-        assert "Error when calling 'get_mass' of a 'BHTree', errorcode is -1" in str(excinfo.value)
-
-
-def test_test7(bhtree_kg):
-    instance = bhtree_kg
-
-    particles = datamodel.Particles(2)
-    assert len(instance.particles) == 0 # TODO: what are we testing here?
-
-    particles.mass = [15.0, 30.0] | units.kg
-    particles.radius = [10.0, 20.0] | units.m
-    particles.position = [[10.0, 20.0, 30.0], [20.0, 40.0, 60.0]] | units.m
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | units.m / units.s
-
-    instance.particles.add_particles(particles)
-    instance.commit_particles()
-
-    check_equal_units(instance.get_mass(1), 15.0 | units.kg)
-
-    check_equal_with_reltol(instance.get_position(1)[2], 30.0 | units.m)
-
-    assert len(instance.particles) == 2
-    check_equal_with_reltol(instance.particles.mass[1], 30.0 | units.kg)
-    check_equal_with_reltol(instance.particles.position[1][2], 60.0 | units.m)
-
-
-def test_test8(bhtree_kg):
-    instance = bhtree_kg
-
-    # TODO: this logic here is repeated above -> fixture. make "empty" fixture in units, and initialize a fixture with particles"
-    particles = datamodel.Particles(2)
-    #self.assertEqual(len(instance.particles), 0) # TODO: tested above already
-
-    particles.mass = [15.0, 30.0] | units.kg
-    particles.radius = [10.0, 20.0] | units.m
-    particles.position = [[10.0, 20.0, 30.0], [20.0, 40.0, 60.0]] | units.m
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | units.m / units.s
-
-    instance.particles.add_particles(particles)
-    instance.commit_particles()
-
-    instance.particles.mass = [17.0, 33.0] | units.kg
-    check_equal_units(instance.get_mass(1), 17.0 | units.kg)
-
-
-def test_test9(bhtree_test9):
-    instance = bhtree_test9
-
-    zero = 0.0 | nbody_system.length
-    gravity = instance.get_gravity_at_point(zero, 1.0 | nbody_system.length, zero, zero)
-    for f in gravity:
-        check_equal_with_reltol(f, 0.0 | nbody_system.acceleration, 3)
-
-    for x in (0.25, 0.5, 0.75): # TODO: this we should put into a pytest parameterize
-        x0 = x | nbody_system.length
-        x1 = (2.0 - x) | nbody_system.length
-        potential0 = instance.get_potential_at_point(zero, x0, zero, zero)
-        potential1 = instance.get_potential_at_point(zero, x1, zero, zero)
-        fx0, fy0, fz0 = instance.get_gravity_at_point(zero, x0, zero, zero)
-        fx1, fy1, fz1 = instance.get_gravity_at_point(zero, x1, zero, zero)
-
-        check_equal_with_reltol(fy0, 0.0 | nbody_system.acceleration, 3)
-        check_equal_with_reltol(fz0, 0.0 | nbody_system.acceleration, 3)
-        check_equal_with_reltol(fy1, 0.0 | nbody_system.acceleration, 3)
-        check_equal_with_reltol(fz1, 0.0 | nbody_system.acceleration, 3)
-
-        check_equal_with_reltol(fx0, -1.0 * fx1, 5)
-        fx = (-1.0 / (x0**2) + 1.0 / (x1**2)) * (1.0 | nbody_system.length ** 3 / nbody_system.time ** 2)
-        check_equal_with_reltol(fx, fx0, 2)
-        check_equal_with_reltol(potential0, potential1, 5)
-
-
-def test_test10(bhtree_test10):
-    instance = bhtree_test10
-
-    zero = 0.0 | nbody_system.length
-    gravity = instance.get_gravity_at_point(zero, zero, zero, zero)
-    for f in gravity: # TODO: tested above as well
-        check_equal_with_reltol(f, 0.0 | nbody_system.acceleration, 3)
-
-    for position in (0.25, 0.5, 0.75):
-        p0 = position | nbody_system.length
-        p1 = -position | nbody_system.length
-        for i in range(3):
-            args0 = [zero] * 4
-            args1 = [zero] * 4
-            args0[1 + i] = p0
-            args1[1 + i] = p1
-            f0 = instance.get_gravity_at_point(*args0)
-            f1 = instance.get_gravity_at_point(*args1)
-
-            for j in range(3):
-                if j != i:
-                    check_equal_with_reltol(f0[j], 0.0 | nbody_system.acceleration, 3)
-                    check_equal_with_reltol(f1[j], 0.0 | nbody_system.acceleration, 3)
-                else:
-                    check_equal_with_reltol(f0[j], -1.0 * f1[j], 5)
-
-def test_test11(bhtree_kg):
-    instance = bhtree_kg
-
-    particles = datamodel.Particles(2)
-    #self.assertEqual(len(instance.particles), 0) # TODO: tested in other function too
-
-    particles.mass = [15.0, 30.0] | units.kg
-    particles.radius = [10.0, 20.0] | units.m
-    particles.position = [[10.0, 20.0, 30.0], [20.0, 40.0, 60.0]] | units.m
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | units.m / units.s
-
-    instance.particles.add_particles(particles)
-
-    copyof = instance.particles.copy()
-    check_equal_with_reltol(copyof[1].mass, 30 | units.kg, 6)
-
-    copyof[1].mass = 35 | units.kg
-    copyof.copy_values_of_all_attributes_to(instance.particles)
-
-    check_equal_with_reltol(instance.particles[1].mass, 35 | units.kg, 6)
-
-
-def test_test12(bhtree_kg):
-    instance = bhtree_kg
-
-    particles = datamodel.Particles(2)
-    #self.assertEqual(len(instance.particles), 0) # checked above. delete?
-
-    particles.mass = [15.0, 30.0] | units.kg
-    particles.radius = [10.0, 20.0] | units.m
-    particles.position = [[10.0, 20.0, 30.0], [20.0, 40.0, 60.0]] | units.m
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | units.m / units.s
-
-    instance.particles.add_particles(particles)
-    instance.commit_particles()
-
-    copyof = instance.particles.copy()
-
-    instance.set_state(1, 16 | units.kg, 20.0 | units.m, 40.0 | units.m, 60.0 | units.m,
-                             1.0 | units.ms, 1.0 | units.ms, 1.0 | units.ms)
-
-    curr_state = instance.get_state(1)
-    for expected, actual in zip((16 | units.kg, 20.0 | units.m, 40.0 | units.m, 60.0 | units.m,
-                             1.0 | units.ms, 1.0 | units.ms, 1.0 | units.ms, 0 | units.m), curr_state):
-        check_equal_with_reltol(actual, expected)
-
-    # TODO: is this just repeating the code from above
-    instance.set_state(1, 16 | units.kg, 20.0 | units.m, 40.0 | units.m, 60.0 | units.m,
-                             1.0 | units.ms, 1.0 | units.ms, 1.0 | units.ms, 20.0 | units.m)
-    curr_state = instance.get_state(1)
-    for expected, actual in zip((16 | units.kg, 20.0 | units.m, 40.0 | units.m, 60.0 | units.m,
-                             1.0 | units.ms, 1.0 | units.ms, 1.0 | units.ms, 20 | units.m), curr_state):
-        check_equal_with_reltol(actual, expected)
-
-
-def test_test13(bhtree_kg):
-    instance = bhtree_kg
-    particles = datamodel.Particles(2)
-    # self.assertEqual(len(instance.particles), 0) # TODO: repeated. delete?
-
-    particles.mass = [30.0, 30.0] | units.kg
-    particles.radius = [1.0, 1.0] | units.m
-    particles.position = [[-10.0, 0.0, 0.0], [10.0, 0.0, 0.0]] | units.m
-    particles.velocity = [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]] | units.m / units.s
-
-    instance.particles.add_particles(particles)
-    instance.commit_particles()
-
-    #copyof = instance.particles.copy() # not accessed?
-    com = instance.center_of_mass_position
-    check_equal_with_reltol(com[0], quantities.new_quantity(0.0, units.m), constants.precision)
-
-
-def test_bhtree_parameters(bhtree_test14):
-    instance = bhtree_test14
-
-    # TODO: combine the testing of legacy interface and instance.parameters? or parameterize?
-    eps2 = instance.legacy_interface.get_epsilon_squared()
-    assert isinstance(eps2, OrderedDictionary)
-    assert eps2.keys() == ["epsilon_squared", "__result"]
-    assert eps2.values() == [0.125, 0]
-    check_equal_with_reltol(instance.parameters.epsilon_squared, 0.125 | units.AU**2, in_units=units.AU**2)
-
-    for x in [0.01, 0.1, 0.2]:
-        instance.parameters.epsilon_squared = x | units.AU**2
-        check_equal_with_reltol(instance.parameters.epsilon_squared, x | units.AU**2, in_units=units.AU**2)
-
-    time_step = instance.legacy_interface.get_time_step()
-    assert isinstance(time_step, OrderedDictionary)
-    assert time_step.keys() == ["time_step", "__result"]
-    assert time_step.values() == [0.015625, 0]
-
-    check_equal_with_reltol(instance.parameters.timestep, 0.015625 | units.yr, in_units=units.yr)
-
-    for x in [0.001, 0.01, 0.1]:
-        instance.parameters.timestep = x | units.yr
-        check_equal_with_reltol(instance.parameters.timestep, x | units.yr, in_units=units.yr)
-
-    theta = instance.legacy_interface.get_theta_for_tree()
-    assert isinstance(theta, OrderedDictionary)
-    assert theta.keys() == ["theta_for_tree", "__result"]
-    assert theta.values() == [0.75, 0]
-
-    assert instance.parameters.opening_angle == 0.75
-    for x in [0.2, 0.5, 0.7]:
-        instance.parameters.opening_angle = x
-        assert instance.parameters.opening_angle == x, "does not change opening angle parameter"
-
-    gravity = instance.legacy_interface.get_use_self_gravity()
-    assert isinstance(gravity, OrderedDictionary)
-    assert gravity.keys() == ["use_self_gravity", "__result"]
-    assert gravity.values() == [1, 0]
-
-    assert instance.parameters.use_self_gravity == 1
-    for x in [0, 1]:
-        instance.parameters.use_self_gravity = x
-        assert instance.parameters.use_self_gravity == x, "does not change use_self_gravity parameter"
-
-    ncrit = instance.legacy_interface.get_ncrit_for_tree()
-    assert isinstance(ncrit, OrderedDictionary)
-    assert ncrit.keys() == ["ncrit_for_tree", "__result"]
-    assert ncrit.values() == [12, 0]
-
-    assert instance.parameters.ncrit_for_tree == 12
-    for x in [512, 2048, 4096]:
-        instance.parameters.ncrit_for_tree = x
-        assert instance.parameters.ncrit_for_tree == x, "does not change ncrit_for_tree parameter"
-
-    dt_dia = instance.legacy_interface.get_dt_dia()
-    assert isinstance(dt_dia, OrderedDictionary)
-    assert dt_dia.keys() == ["dt_dia", "__result"]
-    assert dt_dia.values() == [1.0, 0]
-
-    check_equal_with_reltol(instance.parameters.dt_dia, 1.0 | units.yr, in_units=units.yr)
-    for x in [0.1, 10.0, 100.0]:
-        instance.parameters.dt_dia = x | units.yr
-        check_equal_with_reltol(instance.parameters.dt_dia, x | units.yr, in_units=units.yr)
-
-def test_effect_of_bhtree_param_epsilon_squared(bhtree_for_epsilon_squared_test):
-    instance, initial_direction, final_direction = bhtree_for_epsilon_squared_test
-
-    # Small values of epsilon_squared should result in normal earth-sun dynamics: rotation of 90 degrees
-    check_equal_with_abstol(abs(final_direction[0]), abs(initial_direction + math.pi/2.0), 2)
-
-    # Large values of epsilon_squared should result in ~ no interaction
-    check_equal_with_abstol(final_direction[-1], initial_direction, 2)
-
-    # Outcome is most sensitive to epsilon_squared when epsilon_squared = d(earth, sun)^2
-    delta = [abs(final_direction[i+1]-final_direction[i]) for i in range(len(final_direction)-1)]
-    assert max(delta) == delta[len(final_direction)//2 - 1]
-
-def test_total_energy(bhtree_to_test_energy): # formerly test16
-    instance = bhtree_to_test_energy
-
-    energy_total_t0 = instance.potential_energy + instance.kinetic_energy
-    request = instance.evolve_model.asynchronous(1.0 | nbody_system.time)
-    request.result()
-    energy_total_t1 = instance.potential_energy + instance.kinetic_energy
-
-    check_equal_with_reltol(energy_total_t0, energy_total_t1, 3)
-
-
-def test_collision_detection(bhtree_for_collision_detection): # formerly test17
-    instance, particles, collisions = bhtree_for_collision_detection
-
-    assert collisions.is_set(), "collisions not set"
-    assert instance.model_time < 0.5 | nbody_system.time, "time too big"
-    assert len(collisions.particles(0)) == 3, "mismatch in N particles"
-    assert len(collisions.particles(1)) == 3, "mismatch in N particles"
-    assert len(particles - collisions.particles(0) - collisions.particles(1)) == 1
-
-    left = abs(collisions.particles(0).x - collisions.particles(1).x)
-    right = collisions.particles(0).radius + collisions.particles(1).radius
-    assert all(left < right) # TODO: useful error message?
-
-    # TODO: this is doing some updates here - put code elsewhere?
-    sticky_merged = datamodel.Particles(len(collisions.particles(0)))
-    sticky_merged.mass = collisions.particles(0).mass + collisions.particles(1).mass
-    sticky_merged.radius = collisions.particles(0).radius
-    for p1, p2, merged in zip(collisions.particles(0), collisions.particles(1), sticky_merged):
-        merged.position = (p1 + p2).center_of_mass()
-        merged.velocity = (p1 + p2).center_of_mass_velocity()
-
-    instance.particles.remove_particles(collisions.particles(0) + collisions.particles(1))
-    instance.particles.add_particles(sticky_merged)
-
-    instance.evolve_model(1.0 | nbody_system.time)
-
-    assert collisions.is_set(), "collisions not set"
-    assert instance.model_time < 1.0 | nbody_system.time, "time too big"
-    assert len(collisions.particles(0)) == 1, "mismatch in N particles"
-    assert len(collisions.particles(1)) == 1, "mismatch in N particles"
-    assert len(instance.particles - collisions.particles(0) - collisions.particles(1)) == 2
-
-    left = abs(collisions.particles(0).x - collisions.particles(1).x)
-    right = collisions.particles(0).radius + collisions.particles(1).radius
-    assert all(left < right) # TODO: useful error message?
-
-
-# TODO: what is the additional benefit of this test? the evolution of the model
-# is also tested elsewhere I think?
-def test_test18(bhtree_test18):
-    instance = bhtree_test18
-    instance.stopping_conditions.number_of_steps_detection.enable()
-    instance.parameters.stopping_conditions_number_of_steps = 2
-    assert instance.parameters.stopping_conditions_number_of_steps == 2
-
-    instance.evolve_model(10 | nbody_system.time)
-    assert instance.stopping_conditions.number_of_steps_detection.is_set()
-    assert instance.model_time < 10 | nbody_system.time
-
-
-
-def test_test19(bhtree_test18):
-    instance = bhtree_test18
-
-    instance.stopping_conditions.timeout_detection.enable()
-
-    very_short_time_to_evolve = 1 | units.s
-    very_long_time_to_evolve = 1e9 | nbody_system.time
-
-    instance.parameters.stopping_conditions_timeout = very_short_time_to_evolve
-
-    assert instance.parameters.stopping_conditions_timeout == very_short_time_to_evolve
-
-    start = time.time()
-    instance.evolve_model(very_long_time_to_evolve)
-    end = time.time()
-
-
-    assert instance.stopping_conditions.timeout_detection.is_set()
-    assert very_short_time_to_evolve.value_in(units.s) + 2 > end - start, "early stopping fails"
-
-    instance.cleanup_code()
-
-def test_test20(bhtree_test18):
-    instance = bhtree_test18
-
-    very_short_time_to_evolve = 1 | units.s
-
-    instance.parameters.stopping_conditions_timeout = very_short_time_to_evolve
-    assert instance.parameters.stopping_conditions_timeout == very_short_time_to_evolve
-
-    codeparticles1 = instance.particles
-    instance.particles.add_particle(datamodel.Particle(
-        position=[0, 1, 2] | nbody_system.length,
-        velocity=[0, 0, 0] | nbody_system.speed,
-        radius=0.005 | nbody_system.length,
-        mass=1 | nbody_system.mass
-    ))
-    codeparticles2 = instance.particles
-    assert codeparticles2 is codeparticles1
-    instance.cleanup_code()
-    codeparticles3 = instance.particles
-    assert codeparticles1 is not codeparticles3, "clean up does not work"
-
-
-def test_test21(bhtree_test18):
-    instance = bhtree_test18
-
-    instance.parameters.epsilon_squared = (1e-5 | nbody_system.length)**2
-    check_equal_with_reltol(instance.potential_energy, -0.1 | nbody_system.energy, 5)
-
-
-def test_test22(bhtree_test18):
-    instance = bhtree_test18
-
-    instance.commit_particles()
-    p = datamodel.Particle(
-        x=1.0 | nbody_system.length,
-        y=2.0 | nbody_system.length,
-        z=3.0 | nbody_system.length,
-        vx=1.0 | nbody_system.speed,
-        vy=2.0 | nbody_system.speed,
-        vz=3.0 | nbody_system.speed,
-        mass=1.0 | nbody_system.mass,
-        radius=4.0 | nbody_system.length,
-    )
-    instance.particles.add_particle(p)
-
-    radii = [x.radius for x in instance.particles]
-    expected = [x | nbody_system.length for x in [0.005, 0.005, 4.0]]
-    assert all([x == y for x,y in zip(radii, expected)]), \
-            "cannot add new particle with different radius"
-
-
-def test_test23(bhtree_test23):
-    instance, particles = bhtree_test23
-    instance.evolve_model(0.1 | nbody_system.time)
-    assert instance.particles[0].vy <= 0 | nbody_system.speed
-
-    check_equal_with_reltol(instance.particles[0].x, 0.1 | nbody_system.length, 4)
-
-    instance.particles.new_channel_to(particles).copy()
-    particles.vy = 1 | nbody_system.speed
-    particles.new_channel_to(instance.particles).copy()
-
-    instance.evolve_model(0.2 | nbody_system.time)
-
-    assert instance.particles[0].vy > 0 | nbody_system.speed
-    check_equal_with_reltol(instance.particles[0].y, 0.1 | nbody_system.length, 4)
-
-
-
-# NOTES
-# test2 seems mostly for the plotting? there's a test for the sun radius,
-# but this just tests that attributes of the same instance (sun) is the same?
-# test3 also does not test anything
-# fct new_system_of_sun_and_earth not used anywhere
-# what is the difference between test4 and 5?
-# TODO: after all is cleaned up, try to refactor the fixtures and also
-# the tests themselves where necessary
-# also: maybe it makes sense to separate logic for setup and logic for testing,
-# as I do here? and then reconsider refactoring both separately
-# TODO: turn off printing of sources when running pytest
-
-
